+++ packages/plugin-grid/src/index.ts
@@ -1,1 +1,0 @@
-// This plugin will add support for fragments, 1 / 3 and such values